// src/shared/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const StatusOS = {
  RECEBIDA: 'RECEBIDA',
  EM_DIAGNOSTICO: 'EM_DIAGNOSTICO',
  AGUARDANDO_APROVACAO: 'AGUARDANDO_APROVACAO',
  EM_EXECUCAO: 'EM_EXECUCAO',
  FINALIZADA: 'FINALIZADA',
  ENTREGUE: 'ENTREGUE'
} as const

export type StatusOS = (typeof StatusOS)[keyof typeof StatusOS]


export const StatusOrcamento = {
  AGUARDANDO_APROVACAO: 'AGUARDANDO_APROVACAO',
  APROVADO: 'APROVADO',
  REJEITADO: 'REJEITADO'
} as const

export type StatusOrcamento = (typeof StatusOrcamento)[keyof typeof StatusOrcamento]


export const TipoMovimentacaoEstoque = {
  ENTRADA: 'ENTRADA',
  BAIXA: 'BAIXA'
} as const

export type TipoMovimentacaoEstoque = (typeof TipoMovimentacaoEstoque)[keyof typeof TipoMovimentacaoEstoque]


export const PerfilUsuario = {
  ADMINISTRADOR: 'ADMINISTRADOR',
  ATENDENTE: 'ATENDENTE',
  MECANICO: 'MECANICO',
  ALMOXARIFE: 'ALMOXARIFE'
} as const

export type PerfilUsuario = (typeof PerfilUsuario)[keyof typeof PerfilUsuario]